[FIX] mail: slight discuss sidebar visual improvements
1. Discuss search in discuss sidebar is slightly off-aligned with
discuss content header thread avatar.
2. Discuss search input and meeting buttons are too squarish and not
cohesive enough, i.e too much gap.
3. Search button in compact mode looks too different from the fake
input in expanded mode.
4. There's too much spacing between mailboxes and discuss search
part in compact mode.
5. Mailboxes items in compact mode looks too small compared to other
items

This commit fixes each problem as follow:
1. Discuss search top alignment reduced to align with thread avatar.
2. Rounder search and new meeting UI elements.
3. Search button in compact mode also uses fake input visual but
shows only icon without the "search conversations..." placeholder.
4. Reduced bottom padding of discuss search in sidebar in compact
mode.
5. Mailboxes use large icon in compact mode.
