+++ addons/mail/static/src/core/public_web/discuss_search.js
@@ -2,6 +2,7 @@ import { useHover } from "@mail/utils/common/hooks";
 import { Component } from "@odoo/owl";
 import { Dropdown } from "@web/core/dropdown/dropdown";
 import { useDropdownState } from "@web/core/dropdown/dropdown_hooks";
+import { _t } from "@web/core/l10n/translation";
 import { useService } from "@web/core/utils/hooks";
 
 /**
@@ -30,7 +31,20 @@ export class DiscussSearch extends Component {
                 }
             },
         });
+        this.meetingHover = useHover(["meeting-btn", "meeting-floating"], {
+            onHover: () => {
+                if (this.store.discuss.isSidebarCompact) {
+                    this.meetingFloating.isOpen = true;
+                }
+            },
+            onAway: () => {
+                if (this.store.discuss.isSidebarCompact) {
+                    this.meetingFloating.isOpen = false;
+                }
+            },
+        });
         this.searchFloating = useDropdownState();
+        this.meetingFloating = useDropdownState();
     }
 
     get class() {
@@ -41,6 +55,17 @@ export class DiscussSearch extends Component {
                 .join(" ");
         }
         return this.props.class;
+    }
+
+    get newMeetingText() {
+        return _t("New Meeting");
+    }
+
+    onClickNewMeeting() {
+        this.store.startMeeting();
+        if (this.env.inMessagingMenu) {
+            this.env.inMessagingMenu.dropdown.close();
+        }
     }
 
     onClickSearchConversations() {
