+++ addons/mail/static/src/core/public_web/discuss_sidebar.xml
@@ -6,10 +6,11 @@
                 <div class="o-mail-DiscussSidebar d-flex d-print-none flex-column overflow-auto o-scrollbar-thin flex-shrink-0 h-100 border-end border-secondary z-1 o-mail-discussSidebarBgColor o-gap-0_5" t-att-class="{ 'o-compact': store.discuss.isSidebarCompact }">
                     <DiscussSearch class="{
                         'position-sticky top-0 z-2 o-mail-discussSidebarBgColor': true,
-                        'py-2': !store.inPublicPage,
+                        'pt-2 pb-1': !store.inPublicPage and store.discuss.isSidebarCompact,
+                        'py-2': !store.inPublicPage and !store.discuss.isSidebarCompact,
                         'pt-2 o-px-2_5': store.inPublicPage,
                         'flex-column': store.discuss.isSidebarCompact and !ui.isSmall,
-                        'o-ps-2_5 o-pe-1_5 mt-1': !store.discuss.isSidebarCompact or ui.isSmall,
+                        'o-px-2_5 o-mt-0_5': !store.discuss.isSidebarCompact or ui.isSmall,
                     }"/>
                     <t t-foreach="discussSidebarItems" t-as="item" t-key="item_index" t-component="item"/>
                 </div>
