--- addons/mail/static/src/core/web/discuss_search_patch.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<templates xml:space="preserve">
-    <t t-inherit="mail.DiscussSearch" t-inherit-mode="extension">
-        <xpath expr="//*[@t-ref='root']" position="inside">
-            <Dropdown t-if="store.discuss.isSidebarCompact and !ui.isSmall" state="meetingFloating" position="'right-start'" menuClass="'o-mail-DiscussSidebar-floatingMenu o-discuss-dropdownMenu border-secondary px-2 mx-2 my-0 min-w-0'" manual="true">
-                <t t-call="mail.DiscussSearch.newMeeting"/>
-                <t t-set-slot="content">
-                    <div t-ref="meeting-floating">
-                        <span class="text-muted user-select-none" t-esc="newMeetingText"/>
-                    </div>
-                </t>
-            </Dropdown>
-            <t t-else="" t-call="mail.DiscussSearch.newMeeting"/>
-        </xpath>
-    </t>
-
-    <t t-name="mail.DiscussSearch.newMeeting">
-        <button class="btn btn-primary rounded d-flex align-items-center gap-1" t-attf-class="{{ ui.isSmall ? 'p-2' : 'o-p-1_5' }}" t-att-title="newMeetingText" t-on-click="onClickNewMeeting" data-hotkey="m" t-ref="meeting-btn"><i class="fa fa-fw fa-user-plus opacity-75"/></button>
-    </t>
-</templates>
