--- addons/mail/static/src/core/web/discuss_sidebar_mailboxes.scss
@@ -1,4 +1,0 @@
-.o-mail-Mailbox.o-compact {
-    padding-top: map-get($spacers, 2) + map-get($spacers, 1) / 2; // so that button is square like other items
-    padding-bottom: map-get($spacers, 2) + map-get($spacers, 1) / 2; // so that buton is square like other items
-}
