+++ addons/mail/static/src/discuss/core/public_web/discuss_sidebar_categories.scss
@@ -80,6 +80,8 @@ $o-mail-DiscussSidebarChannel-borderOpacity: .05;
     left: 0px;
 
     .o-mail-DiscussSidebarChannel-container.o-bordered & {
+        left: -1px; // compensate for border
+
         &.o-mail-DiscussSidebarChannel, &.o-compact {
             filter: drop-shadow(1px 2px 0px $white) drop-shadow(-1px 1px 0px $white) drop-shadow(1px -1px 0px $white) drop-shadow(-1px -2px 0px $white);
         }
